src/requirements.txt
pip
Django
nodeenv
gunicorn
python-dotenv
biopython
pandas
numpy
XlsxWriter
psycopg2
livereload
